Initial Work on Instruction Decode Logic
This change adds some initial logic for decoding CPU instructions. It
can currently decode instruction format type and all fields from any of
those types. Some of the logic currently uses registers though and
should probably be done with wires instead so it can be accessed during
the decode stage.

This change also removes the hardcoded halting logic in favor of logic
that halts the CPU when an invalid instruction is encountered. This
currently correctly stops the CPU once it runs off the end of the
instruction buffer so it works for now...