Added Window Resize Functionality
The Vulkan swapchain and associated resources are now updated in
response to window resize events. This should allow the application to
render correctly at whatever resolution the user desires.

There is still an issue related to handling window minimize events
which is currently causing crashes. This will be fixed in a later
change.