Simulation Logic Cleanup + Drag-n-Drop ELF Loading
This change moves all of the code that interacts with the devsim device
into a Simulation object. It also exposes reset, pause, and resume
functionality for the simulator through the GUI. As of this change, ELF
files are now loaded into the simulator by dragging and dropping them
onto the window.